+++ hasoffer.core/src/main/java/hasoffer/core/bo/system/SearchLogBo.java
@@ -10,13 +10,13 @@ public class SearchLogBo {
     String site;
     long productId;
     long cmpSkuId;
-    double price;
+    float price;
     long category;
     String brand;
     String sourceId;
 
     public SearchLogBo(String sourceId, String keyword, String brand, String site,
-                       long category, long productId, long cmpSkuId, double price) {
+                       long category, long productId, long cmpSkuId, float price) {
         this.sourceId = sourceId;
         this.keyword = keyword;
         this.site = site;
@@ -75,11 +75,11 @@ public class SearchLogBo {
         this.productId = productId;
     }
 
-    public double getPrice() {
+    public float getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(float price) {
         this.price = price;
     }
